refactor: Clarify names
